+++ packages/avatar/src/index.js
@@ -1,86 +1,139 @@
-import React from "react"
-import styled, { css } from "styled-components"
-import { space, layout, flexbox } from "styled-system"
+import React, { forwardRef } from "react"
+import { Box } from "@rent_avail/layout"
 
-const smallStyles = css`
-  width: 4rem;
-  padding-right: initial;
-  > *:not(.initials) {
-    display: none;
-  }
-`
+// .item {
+//   width: 3rem;
+//   height:3rem;
+//   border-radius: 50%;
+//   background: black;
+//   border: 4px solid transparent;
+// }
 
-const StyledAvatar = styled.div.attrs({
-  role: "user-chip"
-})`
-  position: relative;
-  border-radius: 2rem;
-  background: ${({ theme }) => theme.colors.ui_300};
-  display: inline-flex;
-  align-items: center;
-  height: 4rem;
-  padding-right: 2rem;
-  ${space};
-  ${layout};
-  ${flexbox};
-  img {
-    display: block;
-    width: 3rem;
-    height: 3rem;
-    border-radius: 50%;
-    object-fit: cover;
-    margin-left: 0.5rem;
-  }
-  .name {
-    margin-left: 1rem;
-    overflow: hidden;
-    max-width: 20rem;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    line-height: 24px;
-  }
-  .initials {
-    text-align: center;
-    width: 100%;
-  }
-  .icon {
-    margin-left: 2rem;
-  }
-  ${({ small }) => small && smallStyles}
-`
+// .group {
+//   display: inline-flex;
+//   .item {
+//     border-color: var(--bg-color);
+//   }
+//   .item:not(:last-of-type) {
+//      margin-right: -1rem;
+//   }
+// }
 
-export function Avatar({
-  small,
-  name,
-  email,
-  photoUrl,
-  initials,
-  icon: Icon,
-  ...props
-}) {
-  if (small) {
-    return (
-      <StyledAvatar small {...props}>
-        <p className="initials">{initials.substring(0, 2)}</p>
-      </StyledAvatar>
-    )
-  }
+function AvatarGroup({ sx = {}, ...props }) {
   return (
-    <StyledAvatar {...props}>
-      <img
-        src={photoUrl}
-        alt={`${name}'s profile shot`}
-        className="profile-photo"
-      />
-      <span className="name">{name || email}</span>
-      {Icon && <Icon className="icon" />}
-    </StyledAvatar>
+    <Box
+      {...props}
+      sx={{
+        ...sx,
+        display: "flex",
+        "& > *": { borderColor: "" },
+        "& > *:not(:first-of-type)": { ml: "-1rem" },
+      }}
+    />
   )
 }
 
-export const AvatarGroup = styled.div`
-  display: flex;
-  > *:not(:first-of-type) {
-    margin-left: -1rem;
+const Avatar = forwardRef(function Avatar(
+  {
+    photo = "https://rentalutions-assets.s3.amazonaws.com/avatars/not-found.svg",
+    name,
+    email,
+    size = "large",
+    initials,
+    icon,
+    sx = {},
+    ...props
+  },
+  ref
+) {
+  if (size === "large") {
+    return (
+      <Box
+        {...props}
+        ref={ref}
+        role="figure"
+        sx={{
+          display: "inline-flex",
+          gap: "1rem",
+          bg: "ui_300",
+          p: "0.5rem",
+          pr: "2rem",
+          borderRadius: "4rem",
+          maxWidth: "25rem",
+          ...sx,
+        }}
+      >
+        <Box
+          as="img"
+          src={photo}
+          sx={{
+            width: "4rem",
+            height: "4rem",
+            borderRadius: "2rem",
+            flex: "0 0 auto",
+            objectFit: "cover",
+            bg: "ui_500",
+          }}
+        />
+        <Box
+          sx={{
+            display: "flex",
+            flexDirection: "column",
+            justifyContent: "center",
+          }}
+        >
+          <Box
+            as="span"
+            className="avatar__name"
+            sx={{
+              overflow: "hidden",
+              overflowWrap: "anywhere",
+              wordBreak: email ? "break-all" : "normal",
+              // This is kind of a hack but works in every major browser except IE
+              display: "-webkit-box",
+              "-webkit-line-clamp": email ? "1" : "2",
+              "-webkit-box-orient": "vertical",
+            }}
+          >
+            {name}
+          </Box>
+          {email && (
+            <Box
+              as="span"
+              className="avatar__email"
+              sx={{
+                overflow: "hidden",
+                color: "ui_700",
+                textOverflow: "ellipsis",
+                whiteSpace: "nowrap",
+              }}
+            >
+              {email}
+            </Box>
+          )}
+        </Box>
+        {icon && <Box as={icon} />}
+      </Box>
+    )
   }
-`
+  return (
+    <Box
+      {...props}
+      ref={ref}
+      role="figure"
+      sx={{
+        display: "flex",
+        textAlign: "center",
+        width: "4rem",
+        borderRadius: "4rem",
+        bg: "ui_300",
+        p: "1rem",
+        ...sx,
+      }}
+    >
+      {initials}
+    </Box>
+  )
+})
+
+export { Avatar, AvatarGroup }